refactor : User -> Member #4
